feat: refactor Navbar and Layout
