+++ components/Layout/Layout.tsx
@@ -1,24 +1,46 @@
-import Head from 'next/head'
-import React, { FC } from 'react'
+import { Navbar } from "components";
+import UnauthorizedScreen from "components/UnauthorizedScreen";
+import Head from "next/head";
+import React, { FC, useEffect, useState } from "react";
+import { useAccount } from "wagmi";
 
 interface Props {
-  title: string
-  pageDescription: string
-  children: React.ReactNode
+  title: string;
+  pageDescription: string;
+  children: React.ReactNode;
 }
 
 export const Layout: FC<Props> = ({ children, title, pageDescription }) => {
+  const { isConnected } = useAccount();
+  const [hydrationLoading, sethydrationLoading] = useState(true);
+  useEffect(() => {
+    sethydrationLoading(false);
+  }, []);
+
+  if (hydrationLoading) {
+    return <h1>Loading...</h1>;
+  }
+
   return (
     <>
-    <Head>
-      <title>{title}</title>
-      <meta name="description" content={pageDescription} />
-      <meta name="og:title" content={pageDescription} />
-    </Head>
-    <nav>
-    
-    </nav>
-    <main>{children}</main>
-  </>
-  )
-}
+      <Head>
+        <title>{title}</title>
+        <meta name="description" content={pageDescription} />
+        <meta name="og:title" content={pageDescription} />
+      </Head>
+      {isConnected ? (
+        <>
+          {" "}
+          <nav>
+            <Navbar />
+          </nav>
+          <main>{children}</main>
+        </>
+      ) : (
+        <UnauthorizedScreen />
+      )}
+      {/* Replace this line to ONLY use Lens Login */}
+      {/* {profile ? <Nav /> : <UnauthorizedScreen />} */}
+    </>
+  );
+};
