+++ components/index.tsx
@@ -1,1 +1,2 @@
 export * from './Layout'
+export * from './Navbar'
