+++ pages/organizations/index.tsx
@@ -1,0 +1,13 @@
+import { Layout } from "components";
+import { NextPage } from "next";
+
+const Organizations: NextPage = () => {
+return (
+   <Layout title="Lenstags | Organizations" pageDescription="TODO: Descripcion de la pagina">
+    <h1>Organizations</h1>
+   </Layout>
+)
+
+}
+
+export default Organizations
